+++ Tests/LinuxMain.swift
@@ -829,6 +829,12 @@ extension PrivateUnitTestRuleTests {
     ]
 }
 
+extension ProhibitedInterfaceBuilderRuleTests {
+    static var allTests: [(String, (ProhibitedInterfaceBuilderRuleTests) -> () throws -> Void)] = [
+        ("testWithDefaultConfiguration", testWithDefaultConfiguration)
+    ]
+}
+
 extension ProhibitedSuperRuleTests {
     static var allTests: [(String, (ProhibitedSuperRuleTests) -> () throws -> Void)] = [
         ("testWithDefaultConfiguration", testWithDefaultConfiguration)
@@ -1340,6 +1346,7 @@ XCTMain([
     testCase(PrivateOutletRuleTests.allTests),
     testCase(PrivateOverFilePrivateRuleTests.allTests),
     testCase(PrivateUnitTestRuleTests.allTests),
+    testCase(ProhibitedInterfaceBuilderRuleTests.allTests),
     testCase(ProhibitedSuperRuleTests.allTests),
     testCase(ProtocolPropertyAccessorsOrderRuleTests.allTests),
     testCase(QuickDiscouragedCallRuleTests.allTests),