+++ CHANGELOG.md
@@ -19,6 +19,11 @@
   [Yusuke Ohashi](https://github.com/junkpiano)
   [#2350](https://github.com/realm/SwiftLint/issues/2350)
 
+* Add `prohibited_interface_builder` opt-in rule to validate that `@IBOutlet`s
+  and `@IBAction`s are not used.  
+  [Marcelo Fabri](https://github.com/marcelofabri)
+  [#2365](https://github.com/realm/SwiftLint/pull/2365)
+
 #### Bug Fixes
 
 * Fix `comma` rule false positives on object literals (for example, images).